vm: fix constantinople rules vm: move HF-selection logic for SSTORE gas into functions
vm: remove HF selection logic
