+++ packages/vm/lib/evm/opcodes/EIP1283.ts
@@ -9,51 +9,49 @@ import { RunState } from './../interpreter'
  * @param {Buffer}   value
  */
 export function updateSstoreGasEIP1283(runState: RunState, found: any, value: Buffer) {
-  if (runState._common.hardfork() === 'constantinople') {
-    const original = found.original
-    const current = found.current
-    if (current.equals(value)) {
-      // If current value equals new value (this is a no-op), 200 gas is deducted.
-      runState.eei.useGas(new BN(runState._common.param('gasPrices', 'netSstoreNoopGas')))
-      return
-    }
-    // If current value does not equal new value
-    if (original.equals(current)) {
-      // If original value equals current value (this storage slot has not been changed by the current execution context)
-      if (original.length === 0) {
-        // If original value is 0, 20000 gas is deducted.
-        return runState.eei.useGas(new BN(runState._common.param('gasPrices', 'netSstoreInitGas')))
-      }
-      if (value.length === 0) {
-        // If new value is 0, add 15000 gas to refund counter.
-        runState.eei.refundGas(new BN(runState._common.param('gasPrices', 'netSstoreClearRefund')))
-      }
-      // Otherwise, 5000 gas is deducted.
-      return runState.eei.useGas(new BN(runState._common.param('gasPrices', 'netSstoreCleanGas')))
-    }
-    // If original value does not equal current value (this storage slot is dirty), 200 gas is deducted. Apply both of the following clauses.
-    if (original.length !== 0) {
-      // If original value is not 0
-      if (current.length === 0) {
-        // If current value is 0 (also means that new value is not 0), remove 15000 gas from refund counter. We can prove that refund counter will never go below 0.
-        runState.eei.subRefund(new BN(runState._common.param('gasPrices', 'netSstoreClearRefund')))
-      } else if (value.length === 0) {
-        // If new value is 0 (also means that current value is not 0), add 15000 gas to refund counter.
-        runState.eei.refundGas(new BN(runState._common.param('gasPrices', 'netSstoreClearRefund')))
-      }
-    }
-    if (original.equals(value)) {
-      // If original value equals new value (this storage slot is reset)
-      if (original.length === 0) {
-        // If original value is 0, add 19800 gas to refund counter.
-        runState.eei.refundGas(
-          new BN(runState._common.param('gasPrices', 'netSstoreResetClearRefund'))
-        )
-      } else {
-        // Otherwise, add 4800 gas to refund counter.
-        runState.eei.refundGas(new BN(runState._common.param('gasPrices', 'netSstoreResetRefund')))
-      }
-    }
-    return runState.eei.useGas(new BN(runState._common.param('gasPrices', 'netSstoreDirtyGas')))
+  const original = found.original
+  const current = found.current
+  if (current.equals(value)) {
+    // If current value equals new value (this is a no-op), 200 gas is deducted.
+    runState.eei.useGas(new BN(runState._common.param('gasPrices', 'netSstoreNoopGas')))
+    return
   }
+  // If current value does not equal new value
+  if (original.equals(current)) {
+    // If original value equals current value (this storage slot has not been changed by the current execution context)
+    if (original.length === 0) {
+      // If original value is 0, 20000 gas is deducted.
+      return runState.eei.useGas(new BN(runState._common.param('gasPrices', 'netSstoreInitGas')))
+    }
+    if (value.length === 0) {
+      // If new value is 0, add 15000 gas to refund counter.
+      runState.eei.refundGas(new BN(runState._common.param('gasPrices', 'netSstoreClearRefund')))
+    }
+    // Otherwise, 5000 gas is deducted.
+    return runState.eei.useGas(new BN(runState._common.param('gasPrices', 'netSstoreCleanGas')))
+  }
+  // If original value does not equal current value (this storage slot is dirty), 200 gas is deducted. Apply both of the following clauses.
+  if (original.length !== 0) {
+    // If original value is not 0
+    if (current.length === 0) {
+      // If current value is 0 (also means that new value is not 0), remove 15000 gas from refund counter. We can prove that refund counter will never go below 0.
+      runState.eei.subRefund(new BN(runState._common.param('gasPrices', 'netSstoreClearRefund')))
+    } else if (value.length === 0) {
+      // If new value is 0 (also means that current value is not 0), add 15000 gas to refund counter.
+      runState.eei.refundGas(new BN(runState._common.param('gasPrices', 'netSstoreClearRefund')))
+    }
+  }
+  if (original.equals(value)) {
+    // If original value equals new value (this storage slot is reset)
+    if (original.length === 0) {
+      // If original value is 0, add 19800 gas to refund counter.
+      runState.eei.refundGas(
+        new BN(runState._common.param('gasPrices', 'netSstoreResetClearRefund'))
+      )
+    } else {
+      // Otherwise, add 4800 gas to refund counter.
+      runState.eei.refundGas(new BN(runState._common.param('gasPrices', 'netSstoreResetRefund')))
+    }
+  }
+  return runState.eei.useGas(new BN(runState._common.param('gasPrices', 'netSstoreDirtyGas')))
 }
